backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.3.2

Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2

idna==3.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
python-dotenv==1.0.1

